% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pma_wrappers.R
\name{as.cv.glmnet}
\alias{as.cv.glmnet}
\title{Convert an object to cv.glmnet}
\usage{
as.cv.glmnet(x, ...)
}
\arguments{
\item{x}{An object for which a method exists.}

\item{...}{Arguments passed to or from other methods.}
}
\value{
An object of class \code{cv.glmnet},
as documented in \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}.
}
\description{
Create a \code{cv.glmnet} object from an object for which a method exists,
so that all methods for \code{cv.glmnet} objects can be used.
}
\examples{
x <- "a"
converted <- as.cv.glmnet(x)
}
