% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonapersona.R
\docType{data}
\name{bonapersona}
\alias{bonapersona}
\title{Data from 'The behavioral phenotype of early life adversity'}
\format{
A data.frame with 734 rows and 65 columns.
}
\usage{
data(bonapersona)
}
\description{
This meta-analysis of rodent studies examined whether early life adversity
(ELA) alters cognitive performance in several domains.
The data include over 400 independent experiments, involving approximately
8600 animals.
}
\references{
Bonapersona, V., Kentrop, J., Van Lissa, C. J., van der Veen,
R., Joels, M., & Sarabdjitsingh, R. A. (2019).
The behavioral phenotype of early life adversity: A 3-level meta-analysis of
rodent studies. Neuroscience & Biobehavioral Reviews, 102, 299–307.
\doi{10.1016/j.neubiorev.2019.04.021}
}
\keyword{datasets}
