% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brma.R
\name{brma}
\alias{brma}
\title{Conduct Bayesian Regularized Meta-Analysis}
\usage{
brma(
  formula,
  data,
  vi = "vi",
  study = "study",
  method = "hs",
  scale = TRUE,
  prior = switch(method, lasso = c(lasso_df = 1, lasso_scale = 1), hs = c(hs_df = 1,
    hs_df_global = 1, hs_df_slab = 4, hs_scale_global = 1, hs_scale_slab = 1)),
  ...
)
}
\arguments{
\item{formula}{An object of class \code{formula} (or one that can be coerced to
that class), see \code{\link[stats]{lm}}.}

\item{data}{Optional data.frame containing the variables in the model, see
\code{\link[stats]{lm}}.}

\item{vi}{Character. Name of the column in the \code{data} that
contains the variances of the effect sizes. This column will be removed from
the data prior to analysis. Defaults to \code{"vi"}.}

\item{study}{Character. Name of the column in the
\code{data} that contains the study id. Use this when the data includes
multiple effect sizes per study. This column can be a vector of integers, or
a factor. This column will be removed from the data prior to analysis.
See \code{Details} for more information about analyzing dependent data.}

\item{method}{Character, indicating the type of regularizing prior to use.
Supports one of \code{c("lasso", "hs")}, see Details. Defaults to
\code{"lasso"}.}

\item{scale}{Logical, indicating whether or not to scale the predictors
(defaults to \code{TRUE}, which is recommended so that shrinking affects all
parameters similarly.}

\item{prior}{Numeric vector, specifying the prior to use. Note that the
different \code{method}s require this vector to contain specific named
elements.}

\item{...}{Additional arguments passed on to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}.
Use this, e.g., to override default arguments of that function.}
}
\description{
This function uses Bayesian estimation via the \code{stan} function
\code{\link[stan]{sampling}} to fit a meta-analytic mixed-effects model with
moderators. A lasso or horseshoe prior is used to shrink the regression
coefficients of irrrelevant moderators towards zero. See Details.
}
\details{
The following models are available:
\describe{
\item{lasso}{ More info about the lasso model}
\item{hs}{ More info about the horseshoe model}
}
}
\examples{
set.seed(8)
SimulateSMD()
SimulateSMD(k_train = 50, distribution = "bernoulli")
SimulateSMD(distribution = "bernoulli", model = es * x[ ,1] * x[ ,2])
}
