% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_workshop_data.R
\name{check_workshop_data}
\alias{check_workshop_data}
\title{Check Data for BRMA Workshop}
\usage{
check_workshop_data(df)
}
\arguments{
\item{df}{A \code{data.frame}.}
}
\value{
Invisibly returns a logical TRUE/FALSE, indicating whether all checks
have passed.
}
\description{
This function checks that argument \code{df} is a suitable
\code{data.frame} for completing the workshop on Bayesian Regularized
Meta-Regression, and checks that \code{pema} package dependencies are correctly
installed. It suggests how to remedy any failed checks.
}
\examples{
check_workshop_data(fukkink_lont)
}
