% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxap.R
\name{maxap}
\alias{maxap}
\title{Maximum a posteriori parameter estimate}
\usage{
maxap(x, dens = NULL, ...)
}
\arguments{
\item{x}{Numeric vector.}

\item{dens}{Optional object of class \code{density}. Defaults to \code{NULL}.}

\item{...}{Arguments passed to \code{\link[stats]{density}}}
}
\value{
Atomic numeric vector with the maximum a-posteriori estimate of
vector \code{x}.
}
\description{
Find the parameter estimate with the highest posterior
probability density given a vector of samples.
}
\examples{
maxap(c(1,2,3,4,5))
}
