% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pema-package.R
\docType{package}
\name{pema-package}
\alias{pema-package}
\alias{pema}
\title{pema: Conduct penalized meta-regression.}
\description{
Penalized meta-regression shrinks the regression slopes of
irrelevant moderators towards zero.
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.2. https://mc-stan.org
}
