% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pma_wrappers.R
\name{pma}
\alias{pma}
\alias{pma.formula}
\alias{pma.default}
\title{Conduct LASSO penalized Meta-Analysis}
\usage{
pma(x, ...)

\method{pma}{formula}(formula, data, vi = "vi", study = NULL, ntau = 10, ...)

\method{pma}{default}(x, y, vi, study = NULL, standardize, intercept, ntau = 10, ...)
}
\arguments{
\item{x}{An k x m numeric matrix, where k is the number of effect sizes and m
is the number of moderators.}

\item{formula}{An object of class \code{formula} (or one that can be coerced to
that class), see \code{\link[stats]{lm}}.}

\item{data}{Either a \code{data.frame} containing the variables in the model,
see \code{\link[stats]{lm}}, or a \code{list} of multiple imputed \code{data.frame}s,
or an object returned by \code{\link[mice]{mice}}.}

\item{vi}{Character. Name of the column in the \code{data} that
contains the variances of the effect sizes. This column will be removed from
the data prior to analysis. Defaults to \code{"vi"}.}

\item{study}{Character. Name of the column in the
\code{data} that contains the study id. Use this when the data includes
multiple effect sizes per study. This column can be a vector of integers, or
a factor. This column will be removed from the data prior to analysis.
See \code{Details} for more information about analyzing dependent data.}

\item{ntau}{Numeric, the numer of \code{tau2} values to try during
cross-validation.}

\item{y}{A numeric vector of k effect sizes.}

\item{intercept}{Logical, indicating whether or not an intercept should be included
in the model.}
}
\value{
A \code{list} object of class \code{pma}, with the following structure:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  rma          # A random effects meta-analysis of class rma
  lasso        # A LASSO-penalized regression model of class cv.glmnet
  tau2_cv      # The residual heterogeneity tau2 based on cross-validation
)
}\if{html}{\out{</div>}}
}
\description{
This function conducts Bayesian regularized meta-regression (Van Lissa & Van
Erp, 2021). It uses the \code{stan} function
\link[rstan:stanmodel-method-sampling]{rstan::sampling} to fit the model. A lasso or horseshoe prior is used to
shrink the regression coefficients of irrelevant moderators towards zero.
See Details.
}
