% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_prior.R
\name{shiny_prior}
\alias{shiny_prior}
\title{Interactively Sample from the Prior Distribution}
\usage{
shiny_prior()
}
\value{
NULL, function is called for its side-effect of launching a Shiny
app.
}
\description{
Launches a \code{Shiny} app that allows interactive comparison of
different priors for \code{\link{brma}}.
}
\examples{
\dontrun{
shiny_prior()
}
}
