% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pema-package.R
\docType{package}
\name{pema-package}
\alias{pema-package}
\alias{pema}
\title{pema: Conduct penalized meta-regression.}
\description{
Penalized meta-regression shrinks the regression slopes of
irrelevant moderators towards zero (Van Lissa & Van Erp, 2021).
}
\references{
Van Lissa, C. J., van Erp, S., & Clapper, E. B. (2023).
Selecting relevant moderators with Bayesian regularized
meta-regression. Research Synthesis Methods.
\doi{10.31234/osf.io/6phs5}

Stan Development Team (NA). RStan: the R interface to Stan. R package version
2.26.2. \url{https://mc-stan.org}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cjvanlissa/pema}
  \item \url{https://cjvanlissa.github.io/pema/}
}

}
\author{
\strong{Maintainer}: Caspar J van Lissa \email{c.j.vanlissa@tilburguniversity.edu} (\href{https://orcid.org/0000-0002-0808-5024}{ORCID})

Authors:
\itemize{
  \item Sara J van Erp
}

}
